Fixed attempt at deleting extraction folders, Add games-files now prompts with owner rather than primary stage.

+++ Rom-tools/src/main/java/com/djrapitops/rom/backend/processes/FileProcesses.java
@@ -177,7 +177,7 @@ public class FileProcesses {
     public static void cleanFolder(File folder) {
         Verify.isTrue(folder.isDirectory(), () -> new IllegalArgumentException("Not a folder."));
         for (File file : Objects.requireNonNull(folder.listFiles())) {
-            if (ArchiveExtractor.isSupportedArchiveFile(file.getName())) {
+            if (file.isFile() && ArchiveExtractor.isSupportedArchiveFile(file.getName())) {
                 try {
                     Files.deleteIfExists(file.toPath());
                 } catch (IOException e) {
